+++ examples/spot/account/get_margin_account_details.py
@@ -1,0 +1,34 @@
+import logging
+
+from bitmart.api_account import APIAccount
+from bitmart.lib.cloud_exceptions import APIException
+from bitmart.lib.cloud_utils import config_logging
+
+config_logging(logging, logging.DEBUG)
+logger = logging.getLogger(__name__)
+
+accountAPI = APIAccount(api_key="Your_Api_Key",
+                        secret_key="Your_Secret_Key",
+                        memo="Your_Memo",
+                        logger=logger)
+
+try:
+    response = accountAPI.get_margin_account_details_isolated()[0]
+    logger.info(response)
+except APIException as error:
+    logger.error(
+        "Found error. status: {}, error message: {}".format(
+            error.status_code, error.response
+        )
+    )
+
+
+try:
+    response = accountAPI.get_margin_account_details_isolated(symbol='BTC_USDT')[0]
+    logger.info(response)
+except APIException as error:
+    logger.error(
+        "Found error. status: {}, error message: {}".format(
+            error.status_code, error.response
+        )
+    )