+++ tests/test_api_spot.py
@@ -81,7 +81,7 @@ def test_post_batch_orders():
     """Test POST https://api-cloud.bitmart.com/spot/v2/batch_orders"""
     order_params = [{"symbol": "BTC_USDT", "side": "buy", "type": "limit", "size": "0.01", "price": "8800"},
                     {"symbol": "BTC_USDT", "side": "buy", "type": "limit", "size": "0.01", "price": "8800"}]
-    assert spotAPI.post_batch_orders(order_params=order_params)[0]['code'] == 1000
+    assert spotAPI.post_batch_orders(symbol='BTC_USDT', order_params=order_params)[0]['code'] == 1000
 
 
 def test_post_cancel_order_by_orderid():